chore: Move LambdaWrapper to utils
