+++ src/model/encoder_block.py
@@ -2,6 +2,7 @@ import torch
 import torch.nn as nn
 
 from src.model.residual_block import ResidualBlock
+from src.utils.lambda_wrapper import LambdaWrapper
 
 
 class EncoderBlock(nn.Module):
@@ -37,29 +38,3 @@ class EncoderBlock(nn.Module):
         """
         outputs = self.__self_attention_block(inputs, LambdaWrapper(lambda x: self.__self_attn(x, x, x, mask)))
         return self.__feed_forward_block(outputs, self.__feed_forward)
-
-
-class LambdaWrapper(nn.Module):
-    """
-    A PyTorch Module wrapper for lambda functions to enable compatibility with ResidualBlock.
-    """
-
-    def __init__(self, lambda_func):
-        """Initializes the LambdaWrapper with a callable function.
-
-        :param lambda_func: Function to wrap. Must accept a tensor and return a tensor.
-        """
-        super().__init__()
-        self.lambda_func = lambda_func
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Executes the wrapped function on the input tensor.
-
-        :param x: Input tensor of shape (*, d_model) where d_model is the feature dimension.
-
-        :return: Transformed output tensor of same shape as input.
-
-        Note:
-            The input will typically be layer-normalized when used with ResidualBlock.
-        """
-        return self.lambda_func(x)
